+++ app/ExternalFiles/Files.php
@@ -558,6 +558,16 @@ class Files {
 		</p>
 		</div>
 		<ul class="misc-pub-external-file">
+			<?php
+			$date = $external_file_obj->get_date();
+			if( ! empty( $date ) ) {
+			?>
+			<li>
+				<span class="dashicons dashicons-clock"></span> <?php echo __( 'Imported at', 'external-files-in-media-library' ) . ' ' . $date; ?>
+			</li>
+				<?php
+			}
+		?>
 		<li>
 			<?php
 			if ( $external_file_obj->is_available() ) {