+++ app/ExternalFiles/Protocols/Sftp.php
@@ -62,6 +62,15 @@ class Sftp extends Protocol_Base {
 	}
 
 	/**
+	 * Return the title of this protocol object.
+	 *
+	 * @return string
+	 */
+	public function get_title(): string {
+		return __( 'SFTP/SSH', 'external-files-in-media-library' );
+	}
+
+	/**
 	 * Return infos to each given URL.
 	 *
 	 * @return array<int,array<string,mixed>> List of file-infos.